changes to get adult running

+++ utils/training_utils.py
@@ -60,8 +60,8 @@ def parse_arguments():
     parser.add_argument('--ind-method', type=str, choices=['hsic', 'disc'], default='disc')
     parser.add_argument('--ind-method2t', type=str, choices=['hsic', 'none'], default='none')
 
-    parser.add_argument('--zs_frac', type=int, default=0.33)
-    parser.add_argument('--zy_frac', type=int, default=0.33)
+    parser.add_argument('--zs_frac', type=float, default=0.33)
+    parser.add_argument('--zy_frac', type=float, default=0.33)
 
     parser.add_argument('--log_px_weight', type=float, default=1.)
     parser.add_argument('--pred_y_weight', type=float, default=1.)
